tests/wpt/tests/html/semantics/edits/the-ins-element/ins_effect.html

<!DOCTYPE html>
<meta charset=UTF-8>
<title>HTML Test: Text in the ins element should be 'underline'</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#the-ins-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p><ins>underlined text</ins></p>
<div id="log"></div>

<script>
  test(function() {
    var element = document.getElementsByTagName('ins')[0],
        textDecoration = getComputedStyle(element).textDecorationLine ||
                         getComputedStyle(element).textDecoration;
    assert_equals(textDecoration, 'underline');
  });
</script>
